[GPU] Add col_major optional attribute to MMAAttr
For MMA intrinsics with symmetric register layouts we have the freedom
to choose between row major and column major output layouts by simply
swapping the input operands. This adds a flag to `MMAAttr` that allows
specifying whether the result should be column major.

A follow up PR will turn the column major variant on by default for
mfma as this enables vectorization of writes in non-transpose cases.